Desktop/Tic-Tac-Toe-Tovar-branch/Tic-Tac-Toe-Tovar-branch/Tic-Tac-Toe/requirements.txt
# Dependencias para Tic-Tac-Toe con TDD y ORM

# Framework Web
Flask==2.3.3
Flask-CORS==4.0.0

# ORM - SQLAlchemy
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# MySQL Driver
pymysql==1.1.0
cryptography==41.0.3

# Testing - pytest
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0

# Utilidades
python-dotenv==1.0.0